+++ .env
@@ -1,0 +1,10 @@
+PORT=3000
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=12345
+
+require('dotenv').config();
+
+const PORT = process.env.PORT || 3000;
+console.log(`Port: ${PORT}`);
+
