--- frontend/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>To-Do App</title>
-  <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-  <div class="container">
-    <h1>To-Do Application</h1>
-    <form id="task-form">
-      <input type="text" id="task-input" placeholder="Enter a new task" required>
-      <button type="submit">Add Task</button>
-    </form>
-    <ul id="task-list"></ul>
-  </div>
-  <script src="script.js"></script>
-</body>
-</html>
