--- frontend/styles.css
@@ -1,85 +1,0 @@
-body {
-  font-family: Arial, sans-serif;
-  background-color: #f4f4f4;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.container {
-  width: 400px;
-  background: #fff;
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-h1 {
-  text-align: center;
-  color: #333;
-}
-
-form {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
-
-input[type="text"] {
-  width: 70%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-button {
-  width: 25%;
-  background: #007bff;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  padding: 10px;
-  cursor: pointer;
-}
-
-button:hover {
-  background: #0056b3;
-}
-
-ul {
-  list-style: none;
-  padding: 0;
-}
-
-li {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  border-bottom: 1px solid #ddd;
-}
-
-li.completed {
-  text-decoration: line-through;
-  color: #aaa;
-}
-
-li button {
-  margin-left: 10px;
-  background: #dc3545;
-  border: none;
-  color: #fff;
-  border-radius: 4px;
-  cursor: pointer;
-  padding: 5px 10px;
-}
-
-li button.complete-btn {
-  background: #28a745;
-}
-
-li button:hover {
-  opacity: 0.8;
-}
